Refactor root folders, except dashboard.
