+++ app/account/account-form.tsx
@@ -1,8 +1,8 @@
 'use client'
-import { useCallback, useEffect, useState } from 'react'
-import { createClient } from '@/utils/supabase/client'
-import { type User } from '@supabase/supabase-js'
-import Avatar from './avatar'
+import { useCallback, useEffect, useState } from 'react';
+import { createClient } from '@/utils/supabase/client';
+import { type User } from '@supabase/supabase-js';
+import Avatar from './avatar';
 
 export default function AccountForm({ user }: { user: User | null }) {
     const supabase = createClient()
