+++ app/account/avatar.tsx
@@ -1,7 +1,7 @@
 'use client'
-import React, { useEffect, useState } from 'react'
-import { createClient } from '@/utils/supabase/client'
-import Image from 'next/image'
+import React, { useEffect, useState } from 'react';
+import { createClient } from '@/utils/supabase/client';
+import Image from 'next/image';
 
 export default function Avatar({
     uid,
