+++ app/account/page.tsx
@@ -1,5 +1,5 @@
-import AccountForm from './account-form'
-import { createClient } from '@/utils/supabase/server'
+import AccountForm from './account-form';
+import { createClient } from '@/utils/supabase/server';
 
 export default async function Account() {
     const supabase = createClient()
