+++ app/auth/callback/route.ts
@@ -1,7 +1,5 @@
-import { type EmailOtpType } from '@supabase/supabase-js'
-import { type NextRequest, NextResponse } from 'next/server'
-
-import { createClient } from '@/utils/supabase/server'
+import { type NextRequest, NextResponse } from 'next/server';
+import { createClient } from '@/utils/supabase/server';
 
 // Creating a handler to a GET request to route /auth/confirm
 export async function GET(request: NextRequest) {
