+++ app/auth/signout/route.ts
@@ -1,10 +1,9 @@
-import { createClient } from '@/utils/supabase/server'
-import { type NextRequest, NextResponse } from 'next/server'
+import { createClient } from '@/utils/supabase/server';
+import { type NextRequest, NextResponse } from 'next/server';
 
 export async function POST(req: NextRequest) {
     const supabase = createClient()
 
-    // Check if a user's logged in
     const {
         data: { user },
     } = await supabase.auth.getUser()
