+++ app/layout.tsx
@@ -1,13 +1,13 @@
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
-
 import '../shared/styles/global.scss'
+import React from "react";
 
 const Satoshi = localFont({ src: '../public/fonts/satoshi/Satoshi-Medium.otf' })
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Travit Admin',
+  description: 'Travit by Maasivo',
 }
 
 export default function RootLayout({
