--- app/login/actions.ts
@@ -1,23 +1,0 @@
-'use server'
-
-import { redirect } from 'next/navigation'
-import { createClient } from '@/utils/supabase/server'
-
-export async function login(formData: FormData) {
-    const supabase = createClient()
-
-    // type-casting here for convenience
-    // in practice, you should validate your inputs
-    const data = {
-        email: formData.get('email') as string,
-        password: formData.get('password') as string,
-    }
-
-    const { error } = await supabase.auth.signInWithPassword(data)
-
-    if (error) {
-        return redirect("/login?message=Could not authenticate user");
-    }
-
-    redirect('/dashboard')
-}
