+++ app/page.tsx
@@ -1,5 +1,5 @@
 'use client';
 
-export default function Page() {
+function Page() {
   return <></>;
 }
