+++ app/update-password/data/update-password.ts
@@ -22,7 +22,6 @@ export async function updatePassword(formData: FormData) {
     )
     
     if (error) {
-        console.log('ERROR' + error)
         redirect('/error')
     }
